requirements.txt
qbittorrent-api
bencode.py
httpx
pytest-playwright
cf-clearance